+++ NexusSimulatHeure/src/Metier/Source/SourceHeureFin.java
@@ -5,14 +5,14 @@ import Metier.Carte.Point;
 import Metier.DistributionTriangulaire;
 
 public class SourceHeureFin extends Source {
-    private int heureFin;
+    private double heureFin;
     
-    public SourceHeureFin(int heureFin, Point point, int heureDepart, int frequence, DistributionTriangulaire distribution){
+    public SourceHeureFin(double heureFin, Point point, double heureDepart, double frequence, DistributionTriangulaire distribution){
         super(point, heureDepart, frequence, distribution);
         this.heureFin = heureFin;
     }
     
-    public int getheureFin(){
+    public double getheureFin(){
         return heureFin;
     }
     public void setheureFin(int Fin){
